[WWM-102]Add PutResponseBuilder implementation and header files
- Added a new file `PutResponseBuilder.cpp` which contains the implementation of the `PutResponseBuilder` class.
- Added a new file `PutResponseBuilder.h` which contains the declaration of the `PutResponseBuilder` class.

This commit adds the implementation and header files for the `PutResponseBuilder` class, which is responsible for building a response for a PUT request. The `PutResponseBuilder` class implements the `IBuilder` interface and provides methods for updating a file, creating or updating a file, setting the start line and headers of the response, setting the body, resetting the response, and preparing the response.